+++ src/main/java/com/ien/ienapp/control/RolController.java
@@ -32,7 +32,7 @@ public class RolController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Rol> getRolById(@PathVariable Long id) {
+    public ResponseEntity<Rol> getRolById(@PathVariable Integer id) {
         Optional<Rol> rol = rolService.getRolById(id);
         return rol.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
@@ -43,13 +43,13 @@ public class RolController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Rol> updateRol(@PathVariable Long id, @RequestBody Rol rol) {
+    public ResponseEntity<Rol> updateRol(@PathVariable Integer id, @RequestBody Rol rol) {
         Rol updatedRol = rolService.updateRol(id, rol);
         return updatedRol != null ? ResponseEntity.ok(updatedRol) : ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteRol(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteRol(@PathVariable Integer id) {
         rolService.deleteRol(id);
         return ResponseEntity.noContent().build();
     }
