+++ src/main/java/com/ien/ienapp/repository/AlumnoRepository.java
@@ -4,6 +4,6 @@ package com.ien.ienapp.repository;
 import com.ien.ienapp.entity.Alumno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlumnoRepository extends JpaRepository<Alumno, Long> {
+public interface AlumnoRepository extends JpaRepository<Alumno, Integer> {
     // Métodos personalizados si es necesario
 }
