+++ src/main/java/com/ien/ienapp/repository/RolRepository.java
@@ -3,5 +3,5 @@ package com.ien.ienapp.repository;
 import com.ien.ienapp.entity.Rol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolRepository extends JpaRepository<Rol, Long> {
+public interface RolRepository extends JpaRepository<Rol, Integer> {
 }
